+++ backSJreal/database/migrations/2024_05_21_135704_create_hospedajes_table.php
@@ -12,9 +12,9 @@ return new class extends Migration
     public function up(): void
     {
         Schema::create('hospedajes', function (Blueprint $table) {
-            $table->integer('id_hospedaje', true);
-            $table->integer('hospedaje_id_habitacion')->index('fk_habitacion_has_huesped_habitacion1');
-            $table->integer('hospedaje_id_usuario')->index('fk_hospedaje_usuarios1');
+            $table->unsignedBigInteger('id_hospedaje', true);
+            $table->unsignedBigInteger('hospedaje_id_habitacion')->index('fk_habitacion_has_huesped_habitacion1');
+            $table->unsignedBigInteger('hospedaje_id_usuario')->index('fk_hospedaje_usuarios1');
             $table->dateTime('check_in_hospedaje');
             $table->dateTime('check_out_hospedaje');
             $table->integer('cantidad_habitaciones_hospedaje');
